Notification mail service updated

+++ src/main/java/com/benbillion/services/OTPServiceImpl.java
@@ -1,7 +1,6 @@
 package com.benbillion.services;
 
 import com.benbillion.dtos.OTPVerificationRequest;
-import com.benbillion.dtos.ReminderMailRequest;
 import com.benbillion.models.data.Email;
 import com.benbillion.models.data.VerificationOTP;
 import com.benbillion.models.repository.EmailRepository;
@@ -9,6 +8,7 @@ import com.benbillion.models.repository.OTPRepository;
 import exceptions.GenericHandlerException;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.Random;
 
 @Service
@@ -33,21 +33,11 @@ public class OTPServiceImpl implements OTPService {
             stringBuilder.append(num.charAt(random.nextInt(10)));}
         return stringBuilder;
     }
-
-    @Override
-    public String sendOTP(OTPVerificationRequest otpVerificationRequest){
-        return getString(otpVerificationRequest);
-    }
-
-    @Override
-    public String resendOTP(OTPVerificationRequest otpVerificationRequest){
-        otpRepository.deleteAll();
-        return getString(otpVerificationRequest);
-    }
-
     private String getString(OTPVerificationRequest otpVerificationRequest) {
         String generatedOtp = generateOTP().toString();
-        VerificationOTP verificationOTP = new VerificationOTP(generatedOtp);
+        VerificationOTP verificationOTP = new VerificationOTP(generatedOtp,
+                LocalDateTime.now(),
+                LocalDateTime.now().plusMinutes(1));
         otpRepository.save(verificationOTP);
         otpVerificationRequest.setOtp(generatedOtp);
         otpVerificationRequest.setBody("Your One time password is "+generatedOtp
@@ -57,12 +47,36 @@ public class OTPServiceImpl implements OTPService {
     }
 
     @Override
+    public String sendOTP(OTPVerificationRequest otpVerificationRequest){
+        return getString(otpVerificationRequest);
+    }
+    @Override
+    public String resendOTP(OTPVerificationRequest otpVerificationRequest){
+        otpRepository.deleteAll();
+        return getString(otpVerificationRequest);
+    }
+
+    @Override
     public String verifyOTP(OTPVerificationRequest otpVerificationRequest) {
-        VerificationOTP queried = otpRepository.findByOtp(otpVerificationRequest.getOtp())
+        VerificationOTP queriedVerificationOTp = otpRepository.findByOtp(otpVerificationRequest.getOtp())
                 .orElseThrow(()->  new GenericHandlerException("Invalid VerificationOTP"));
+        if(queriedVerificationOTp.getExpiredAt().isBefore(LocalDateTime.now())){
+            throw new GenericHandlerException("Token has been expired");
+        }
+        if(queriedVerificationOTp.getUsedAt()!=null){
+            throw new GenericHandlerException("Otp has already been used");
+        }
+        queriedVerificationOTp.setUsedAt(LocalDateTime.now());
+        otpRepository.save(queriedVerificationOTp);
         Email newEmail = new Email(otpVerificationRequest.getReminderEmail());
-        emailRepository.deleteAll();
+        emailRepository.deleteAll();      // user can only have one emailAdd!
         emailRepository.save(newEmail);  // if verified, save!
         return "Verification Successful";
     }
+
+    public String deleteExpiredOTP(){
+        otpRepository.deleteVerificationOTPByExpiredAtBefore(LocalDateTime.now());
+        return "";
+    }
+
 }
